Implement the chat history feature and fix related bugs
A chat history feature has been introduced in the chat module, which will save conversations between the user and the assistant. Each time a user sends a message, a new entry is added to the chat_history. Also, fix a bug that was caused by not properly handling chat history before. ```